backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RDKit AI Platform - MCP Server for Chemistry"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag2[openai,mcp]>=0.9.3",
    "fastapi>=0.135.1",
    "fastmcp>=0.6.1",
    "mcp[cli]>=1.3.2",
    "openai>=1.59.6",
    "pubchempy>=1.0.4",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "rdkit>=2025.9.6",
    "sse-starlette>=2.2.1",
    "uvicorn>=0.41.0",
]
